Rename utils to index
